Adding lr schedules and changing model's execution
Now there is a function to train and other to
execute the model and an option to train the
model for more than one epoch.

Also, cyclic learning rate is now available.
There is an implementation of CLR at Pytorch,
but is only compatible with optimizers that
has momentum. Since, Adam was the best optimizer,
specially with CLR's, it is good to keep using it.

CLR and one_cycle works with all optimizers.

In the model creating phase it's defined the de-
vice that all the code will run. That avoids
unnecessary if checks in model's loop.