Adding validation steps between epochs
In supervised learning methods it's common to use
a validation set to avoid overfitting.

Neural networks that are builded to performs image
compression commonly dont use validation set for
training.

After some tests, I'd realize that use validation
dont affect the results too much.
That's probably because of the large number
of training examples used, but this feature can be useful
someday.

Here, the model with the lowest loss in validation
set it's the one that will be tested.